+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Added
+
+-   Avatar: Added a `linkProps` and `linkAs` props which will be passed to the Thumbnail child.
+-   UserBlock: Added a `linkProps` and `linkAs` props which will be passed to the Thumbnail child and used to wrap the username.
+
 ## [2.2.0][] - 2022-01-21
 
 ### Changed